adk_imagine/pyproject.toml
[project]
name = "adk-imagine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=0.4.0",
]
